+++ src/components/index.js
@@ -1,4 +1,5 @@
 export { default as BurguerMenu } from 'components/common/BurguerMenu';
 export { default as DoubleArrow } from 'components/common/DoubleArrow';
+export { default as ExpandableSection } from 'components/layouts/ExpandableSection';
 export { default as Menu } from 'components/common/Menu';
 export { default as Navbar } from 'components/common/Navbar';