+++ src/pages/home/projects/ArtsIArcs.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { ExpandableSection } from 'components';
+
+const ArtsIArcs = ({ openTab, ...props }) => {
+    return (
+        <ExpandableSection {...props} id="arts-i-arcs" title="Arts i Arcs" openTab={openTab}>
+            <></>
+        </ExpandableSection>
+    );
+};
+
+export default ArtsIArcs;