+++ src/pages/home/projects/IEMOnline.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { ExpandableSection } from 'components';
+
+const IEMOnline = ({ openTab, ...props }) => {
+    return (
+        <ExpandableSection {...props} id="iem-online" title="IEM Online" isFirstItem openTab={openTab}>
+            <></>
+        </ExpandableSection>
+    );
+};
+
+export default IEMOnline;