+++ src/pages/home/projects/Malamiga.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { ExpandableSection } from 'components';
+
+const Malamiga = ({ openTab, ...props }) => {
+    return (
+        <ExpandableSection {...props} id="malamiga" title="Malamiga" openTab={openTab}>
+            <></>
+        </ExpandableSection>
+    );
+};
+
+export default Malamiga;